% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_active_thread_plot.R
\name{get_active_thread_plot}
\alias{get_active_thread_plot}
\title{This function provide plot active threads during tests.}
\usage{
get_active_thread_plot(input_dt, scenario_name = NULL, interactive = FALSE)
}
\arguments{
\item{input_dt}{input cleaned jmeter log data table}

\item{scenario_name}{add string to highlight scenario name in subtitle.}

\item{interactive}{convert ggplot to interactive plotly.}
}
\description{
This function provide plot active threads during tests.
}
\examples{
get_active_thread_plot(jdt)
}
\keyword{response_time}
\keyword{vizualization}
