% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_jmeter_log.R
\name{get_aggregate_table}
\alias{get_aggregate_table}
\title{A calculate agregated data}
\usage{
get_aggregate_table(input_dt, personal_gpoup_with_load_gpoup = FALSE)
}
\arguments{
\item{input_dt}{input cleaned jmeter log.}

\item{personal_gpoup_with_load_gpoup}{provide aggregated table by grouping with request name and rps group Defaults to FALSE}
}
\description{
This function get aggregated information about performance report.
}
\examples{
get_aggregate_table(jdt)
}
\keyword{aggregating}
\keyword{statistics}
