% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bar_quantile_plot.R
\name{get_bar_quantile_plot}
\alias{get_bar_quantile_plot}
\title{Return bar plot of response times devided by quantiles}
\usage{
get_bar_quantile_plot(input_dt, scenario_name = NULL, interactive = FALSE)
}
\arguments{
\item{input_dt}{input cleaned jmeter log data table}

\item{scenario_name}{add string to highlight scenario name in subtitle.}

\item{interactive}{convert ggplot to interactive plotly.}
}
\description{
This function provide bar plot with response times for labels devided by quantiles.
}
\examples{
get_bar_quantile_plot(jdt)
}
\keyword{response_time}
\keyword{vizualization}
