% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/get_histogram_response_times_distribution_plot.R
\name{get_histogram_response_times_distribution_plot}
\alias{get_histogram_response_times_distribution_plot}
\title{Return distribution of response times}
\usage{
get_histogram_response_times_distribution_plot(
  input_dt,
  scenario_name = NULL,
  group_by_label = TRUE,
  interactive = FALSE
)
}
\arguments{
\item{input_dt}{input cleaned jmeter log data table}

\item{scenario_name}{add string to highlight scenario name in subtitle.}

\item{group_by_label}{split labels by colors.}

\item{interactive}{convert ggplot to interactive plotly.}
}
\description{
This function provide plot with distribution of response times times for labels.
}
\examples{
get_histogram_response_times_distribution_plot(jdt)
}
\keyword{response_time}
\keyword{vizualization}
