% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_response_time_violin_plot_vs_threads.R
\name{get_response_time_violin_plot_vs_threads}
\alias{get_response_time_violin_plot_vs_threads}
\title{Return Violin plots of response times vs active threads}
\usage{
get_response_time_violin_plot_vs_threads(
  input_dt,
  scenario_name = NULL,
  input_request_name = NULL,
  group_by_label = TRUE,
  interactive = FALSE,
  split_by_responseCode = FALSE
)
}
\arguments{
\item{input_dt}{input cleaned jmeter log data table}

\item{scenario_name}{add string to highlight scenario name in subtitle.}

\item{input_request_name}{allow to filter input data for specific request.}

\item{group_by_label}{split labels by colors.}

\item{interactive}{convert ggplot to interactive plotly.}

\item{split_by_responseCode}{allow facet plot by response code.}
}
\description{
This function provide quantile order with response times for labels.
}
\examples{
get_response_time_violin_plot_vs_threads(jdt)
}
\keyword{response_time}
\keyword{vizualization}
