% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_response_times_with_errors_plot.R
\name{get_response_times_with_errors_plot}
\alias{get_response_times_with_errors_plot}
\title{Return plot of response times with errors}
\usage{
get_response_times_with_errors_plot(
  input_dt,
  scenario_name = NULL,
  group_by_label = TRUE,
  interactive = FALSE
)
}
\arguments{
\item{input_dt}{input cleaned jmeter log data table}

\item{scenario_name}{add string to highlight scenario name in subtitle.}

\item{group_by_label}{split labels by colors.}

\item{interactive}{convert ggplot to interactive plotly.}
}
\description{
This function provide plot with response times with errors for labels.
}
\examples{
get_response_times_with_errors_plot(jdt)
}
\keyword{response_time}
\keyword{vizualization}
