% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rps_plot.R
\name{get_rps_plot}
\alias{get_rps_plot}
\title{Return plot of of request per seconds}
\usage{
get_rps_plot(
  input_dt,
  scenario_name = NULL,
  group_by_label = TRUE,
  interactive = FALSE
)
}
\arguments{
\item{input_dt}{input cleaned jmeter log data table}

\item{scenario_name}{add string to highlight scenario name in subtitle.}

\item{group_by_label}{split labels by colors.}

\item{interactive}{convert ggplot to interactive plotly.}
}
\description{
This function provide plot with informmation about request per secind during performance tests.
}
\examples{
get_response_time_plot(jdt)
}
\keyword{per}
\keyword{request}
\keyword{second}
\keyword{vizualization}
