% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_thoghput_per_second_plot.R
\name{get_thoghput_per_second_plot}
\alias{get_thoghput_per_second_plot}
\title{This function provide plot with response times with errors for labels.}
\usage{
get_thoghput_per_second_plot(
  input_dt,
  scenario_name = NULL,
  interactive = FALSE
)
}
\arguments{
\item{input_dt}{input cleaned jmeter log data table}

\item{scenario_name}{add string to highlight scenario name in subtitle.}

\item{interactive}{convert ggplot to interactive plotly.}
}
\description{
This function provide plot with response times with errors for labels.
}
\examples{
get_thoghput_per_second_plot(jdt)
}
\keyword{response_time}
\keyword{vizualization}
